Switch to using strings rather than URIs for Crux ids
This decision is taken to make it easier to add ad-hoc references in
JSON. Currently, Crux attributes whose values are references to Crux entities
must have associated schema transformers in jinx to transform the string to a
URI, prior to submitting into Crux.

This has a potential downside that ad-hoc references are not possible. That is,
references that aren't expected by JSON Schema (additional properties).