Add a mechanism (for overriding the Host header
This is achieved by adding a :juxt.site.alpha/uri-prefix to the config.edn.

If specified, inbound URLs will be uri-prefix + path. If not specified, will
default to concatenating the request's proto (or X-Forwarded-Proto header) and
Host (or X-Forwarded-Host) request header.

This is required to allow Site to operate behind reverse-proxies.