Added juxt.site.alpha/extract-first-projection? query option
This replaces the existing post-processing logic for pull queries whereby a
result set was unnested if the query contains a pull projection as the first project.

The reason for this commit is that the existing logic caused strange results for
queries that have multiple projections (with the first being a pull projection).