Redesign variants to require one fewer entity per variant
The requirement to add a separate 'mapping' entity to join a resource with a
representation feels overly cumbersome for the small number of use-cases it
adds.

This is a BREAKING CHANGE for users of the existing content-negotation
implementation. We've made this change now because it will become much harder to
make these kinds of changes in the future.