Add interrim support for templates to use GraphQL queries
The idea here is that the results that GraphQL queries can produce are much
better for templating. The existing templating uses template models which are
produced from traversing XT entities. This is low-level and not subject to the
authorization treating that GraphQL queries will be subject to. Hence, it's a
potential vulnerability and makes the authoring of templates require a higher
level of authorization than desierable. Rather than ditch the whole templating
story altogether (which may happen in the future), we'll preserve it but on a
GraphQL layer.