Add subQuery @site feature
This allows a value to contain a GraphQL sub-query which can be run by the
field-resolver. The particular query might depend on values in the object-value,
request or resource. For example, a GraphQL query might be run to populate
fields in a document at run-time. The document may need to specify the GraphQL
as part of its data requirements, or the GraphQL query may be user provided.

This opens up a lot of flexible options for utilising GraphQL in the dynamic
systems built on Site.