Promote debug-request, and therefore graphql, into core
This means that request dumps can be served via apex, which allows us to use the
GraphQL query to get the request data and specify a text/html template, which
can be overridden if necessary.

This removes one of the locator strategies, which further simplifies the
code-base. Request dumps are no longer 'special' and can be served using the
same code path as other resources.