Decorate resource state with links
It is useful to decorate the resource state with hyperlinks to other API
operations. This commit extends the OpenAPI implementation to allow these links
to be specified in the OpenAPI description and merged into the resource state.

These are purposely kept out of the GraphQL response, since GraphQL does not
support hypermedia, and this way hypermedia can be grafted on by the REST layer,
and prevents GraphQL being a leaky abstraction. In other words, GraphQL should
not 'know' about or be coupled to hypermedia considerations.