Add login-form module (#59)
* Allow `::site/template-model` to be of type symbol

Backport support for template-model to be a symbol that resolves to a function
that accepts a request map. A template-model that is a map is still accepted
as before. If template-model is neither a symbol nor a map an exception is
thrown.

* Parse redirect URL from _query form value in login-response

Backport change to login-response to extract the redirect URL from _query
form value instead of _redirect. This allows for easier templating as the
list of query parameters can be passed directly to login-response.

* Add login-form opt module

Backport login-form opt module with redirect to login page logic on 401
response. On successful login the user is redirected back to the
requested resource.