Locate resource ahead of authentication
Resources may be partitioned into "protection spaces". A protection space must
be indicated by the resource metadata.

When we authenticate, we need to know the authorization database from which to
look up the user identity.

"A protection space is defined by the canonical root URI (the scheme and
authority components of the effective request URI; see Section 5.5 of [RFC7230])
of the server being accessed, in combination with the realm value if present.
These realms allow the protected resources on a server to be partitioned into a
set of protection spaces, each with its own authentication scheme and/or
authorization database.  The realm value is a string, generally assigned by the
origin server, that can have additional semantics specific to the authentication
scheme."