Change openid_connect to store session in atom
The current implementation of sessions for both Basic and Bearer
authentication stores sessions in an atom. The latest implementation
of openid_connect stores sessions in the database. To unify the
current implementation and allow session ids generated by the OAuth
flow to be used to request new Bearer tokens they are all now stored
in the same place.

Additionally, a fail-to parameter has been added to the login endpoint
to allow the location the user to should be redirected to on login
failure using OAuth to be specified.