Add prompt=login query parameter to authorize call
The prompt query parameter specifies whether the Authorization Server
prompts the End-User for reauthentication. The value login ensures that
the user must reauthenticate and if they fail to do so an error is returned.

This is necessary to allow users to switch between accounts while
authenticating using Auth0.

For further details see: https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest