Change as-hex-str to use apache codec lib
`java.util.HexFormat` was added in Java 17 and therefore by using it we
have a dependency on that version of Java being available wherever the
app is run. To mitigate this dependency org.apache.commons.codec.binary
can be used for converting bytes to hexadecimal strings. It was already
a transative dependency of the app but is now explicitly required.