Add OAuth2 and OpenID Connect Authentication
To enable OpenID Connect to authenticate a user using the
Authorization Code Flow it must perform the following:
* Call the authorize endpoint of a configured OpenID Provider
* Redirect the user to the login prompt (at for example Auth0)
* On successfully redirecting back to the app an Authorization Code
is provided
* Using the Authorization Code the token endpoint is called to
retrieve information about the authenticated user
* This information includes the `iss` (issuer) and `sub` (subject)
which can be used to uniquely identify a user in the system
* A new session is then created for that user and stored in the
session atom used when authenticating by any other means

Other details of the implementation include:
* On returning back to the web app that initiated the OAuth login
process the `return-to` and `fail-to` parameters provided are used
to determine where to redirect the user to

* The authorize endpoint called at the start of the OAuth login
process allows additional parameters to be provided to change the
behaviour of the login flow including `prompt=login`.

The prompt query parameter specifies whether the Authorization Server
prompts the End-User for reauthentication. The value login ensures that
the user must reauthenticate and if they fail to do so an error is returned.
This is necessary to allow users to switch between accounts while
authenticating using Auth0.

For further details see: https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest

* Prepare install-openid functions to accept config

* Change as-hex-str to use apache codec lib

`java.util.HexFormat` was added in Java 17 and therefore by using it we
have a dependency on that version of Java being available wherever the
app is run. To mitigate this dependency org.apache.commons.codec.binary
can be used for converting bytes to hexadecimal strings. It was already
a transative dependency of the app but is now explicitly required.

Co-authored-by: Felix Alm <fel@juxt.pro>