Remove 'action' from signature of 'allowed?' rule
This wasn't paying its weight and hard to justify. There is no reason for a rule
to take the action as a parameter when it already belongs to that action.

The ability to re-use rules and use some attribute of the action 'instance' to
govern behavior feels it would encourage systems that are harder to reason
about.