Remove concepts of 'user-identity' and 'user'
These might be domain concepts, and they may be used in the book, but they
aren't fundamentally Site concepts. They are now not referred to anywhere in
Site's core source code. If examples choose to create such structures, that is
irrelevant to Site.

This requires adding a concept of basic auth credentials, thereby also surfacing
the anomoly that the Alice's form-based logging is matching against her user
identity in basic-auth credentials.