Add diehard dependency for retry of OAuth calls
Occassionally calls to the token endpoint have been terminated by Auth0
so to provide resiliency this change adds the diehard dependency and
uses `with-retry` to catch any exception, log it, and retry twice more
before failing and returning the user to the login screen.

This change also removes the use of a different `fail-to` URL as the
user will always be redirected to the same resource except the `code`
query parameter will only be added if a matched-identity was found.