Add EQL->Datalog ACL-integrated compiler
This is a large commit that represents the results of a research spike.

The `juxt.site.graphql.basic-test` ns demonstrates a new EQL compiler. EQL is
now intended as a canonical query structure that contains everything required to
generate a single XTDB Datalog query.

It is intended that GraphQL be supported by writing a compiler that generates an
equivalent EQL AST. Both the GraphQL query and corresponding SDL would be inputs
to this compiler, since the EQL needs to be annotated with various parameters,
most crucially, the actions that are required to be invoked during field
resolution.

Also in this commit:

* Initial Malli instrumention of some functions in the actions ns.
* Support for 406
* Improved error handling