Add Apache HttpClient and Guava
Add an explicit dependency on Apache HttpClient and Guava for use in URL
parsing and manipulation. These dependencies are currently already used
in the project, but as transitive dependencies of amazonica and auth0
respectively (amongst others).

This change adds a check when importing openid-configuration for AWS
Cognito to replace the authorization endpoint with the logout endpoint
to allow the user to logout and display the universal login page.