Add diehard resilience to OpenID connect auth and AWS Cognito support (#87)
* Add diehard resilience to OpenID connect auth

Add retry behaviour to network calls during OpenID connect
authentication. Also change call to token endpoint to use form
parameters rather than a JSON body as AWS Cognito only supports form
parameters.

* Add Apache HttpClient and Guava

Add an explicit dependency on Apache HttpClient and Guava for use in URL
parsing and manipulation. These dependencies are currently already used
in the project, but as transitive dependencies of amazonica and auth0
respectively (amongst others).

This change adds a check when importing openid-configuration for AWS
Cognito to replace the authorization endpoint with the logout endpoint
to allow the user to logout and display the universal login page.

* Add OAuthCredentials entity for superuser on init

Create entity needed to allow login via AWS Cognito to the superuser account.