Merge pass into site
There doesn't appear to be sufficient value in keeping these namespaces
apart. The benefit of separation does not outweigh the costs of confusion. The
need to know (or check, or guess) whether a given name is in the 'site' or
'pass' namespace adds too much friction. Since none of the names clash (and if
they did, they can be sufficiently qualified) there isn't a need for an entirely
separate namespace.