Replace selected-representation with resource
The purpose of this is to ensure content negotiated resources work the same way
with respect to rules, state and respond.

Before this commit, there was confusion about when to access :juxt.site/resource
and when to access :juxt.site/selected-representation. By removing
:juxt.site/selected-representation, it is now more readable.

The original resource is now linked to using :juxt.site/variant-of