Add resource packages
Resource packages are sets of resources with some package metadata.

Resource packages define resource factories that can create a resource, or
possibly a set of resources matching the same uri-template.

Resource packages can depend on each other.

The aim for resource packages is to allow Site instances to be built from a set
of modules.