Update user and associated roles in OAuth flow (#92)
* Change current user endpoint to use outer join

* Add Rule to init to allow use of /_site/user

Add method that puts a rule into the database to allow any authenticated
user to call the /_site/user endpoint. This endpoint only returns
details about the current user and is nececessary for a user to
determine if they have no roles assigned to them.

* Update user details on OAuth login success

Update or create the necessary User entity on successfully logging in
via OAuth. This allows the user to be created via the OAuth provider and
then immediately used without needing to first create the user in the
database as well. The User entity is created with the id and username
keys set to the name provided by the id_token. The name and email fields
of the User entity are set with the values of the cognito:username and
cognito:groups fields in the id_token respectively.

In addition to creating the User entity, UserRoleMapping entities are
created if a group associated with the user is a valid Role in the
system. Rules for those Roles must be set ahead of time.
UserRoleMappings are updated each time the user logs in so if a user is
removed from a group then the mapping will be removed next time they login.