Fix mail templating of non-strings
clojure.string/replace expects the result of the replacement function to
return a string, or a value that can be cast to a string. To prevent
class cast exceptions where the object cannot be cast to a string str is
called to call .toString() on all returned objects. Specifically, this
allows numeric values to be templated.

Also, change crux aggregate function distinct-string-list to
sorted-string-list. Function must sort the values into a logical order.