Store sessions in Crux instead of atom (#99)
Store user sessions in Crux rather than in an atom to allow for multiple
instances of site to access the same sessions and to persist sessions
between restarts. These changes also update the openid connect tests to
assert that the expected sessions are created in the database.

The expiry time of sessions created by calling the token endpoint has
been extended from 1 hour to 1 week. Sessions created by logging in by
basic auth have been reduced to 1 week from 1 month to be consistent.